rg-146: * Handled requested changes
